+++ src/cli/cli.go
@@ -68,7 +68,14 @@ func preRun(cc *cobra.Command, args []string) error {
 	}
 
 	if !slices.Contains(avoidTheseDescription, cc.Short) {
-		cfg, err := config.GetConfigRepoDetails(ctx, true, nil)
+		overrides := map[string]string{}
+		if cc.Short == "Connect to a S3 repository" {
+			// Get s3 overrides for connect. Ideally we also need this
+			// for init, but we don't reach this block for init.
+			overrides = repo.S3Overrides()
+		}
+
+		cfg, err := config.GetConfigRepoDetails(ctx, true, overrides)
 		if err != nil {
 			log.Error("Error while getting config info to run command: ", cc.Use)
 			return err